signin model
Corrected a syntax issue "items" to "item"
Fixed the issue with not being able to choose size or color.
Fixed links minus/plus/remove button in checkout.
Tried to add a model when user logs out it will take him/her to the same page when loggin in again but i failed so i removed it.
-